% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conflict_location.R
\docType{data}
\name{conflict_locations}
\alias{conflict_locations}
\title{Conflict Locations Data}
\format{
A data frame with 9 rows and 3 variables:
\describe{
\item{location}{A character vector representing the name of the location.}
\item{latitude}{A numeric vector representing the latitude of the location.}
\item{longitude}{A numeric vector representing the longitude of the location.}
}
}
\usage{
conflict_locations
}
\description{
A dataset containing the locations of conflict incidents with their respective latitude and longitude.
}
\examples{
print(conflict_locations)
  plot_conflict_map(conflict_locations)
}
\keyword{datasets}
