% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_conflict_map.R
\name{plot_conflict_map}
\alias{plot_conflict_map}
\title{Plot Conflict Map}
\usage{
plot_conflict_map(df)
}
\arguments{
\item{df}{A data frame containing conflict incident data with latitude and longitude columns.}
}
\value{
A map visualization of conflict incidents.
}
\description{
Plots a map of conflict incidents using OpenStreetMap and leaflet.
}
\examples{
plot_conflict_map(conflict_locations)

}
